Extract Workflow Building Logic into Separate Module
