+++ packages/giselle-engine/src/core/flows/create-and-run-flow.ts
@@ -1,62 +1,15 @@
 import {
-	type FlowTrigger,
 	type FlowTriggerId,
 	type GenerationContextInput,
 	RunId,
 	type Workflow,
-	type Workspace,
 } from "@giselle-sdk/data-type";
-import { buildWorkflowFromNode } from "@giselle-sdk/workflow-utils";
 import type { GiselleEngineContext } from "../types";
-import { getWorkspace } from "../workspaces/utils";
+import { buildWorkflowFromTrigger } from "./build-workflow-from-trigger";
 import { createRun } from "./create-run";
 import { FlowRunIndexObject } from "./run/object";
 import { flowRunPath, workspaceFlowRunPath } from "./run/paths";
 import { type RunFlowCallbacks, runFlow } from "./run-flow";
-import { getFlowTrigger } from "./utils";
-
-async function buildWorkflowFromTrigger(args: {
-	triggerId: FlowTriggerId;
-	context: GiselleEngineContext;
-	useExperimentalStorage: boolean;
-}): Promise<{
-	workflow: Workflow;
-	trigger: FlowTrigger;
-	workspace: Workspace;
-} | null> {
-	const trigger = await getFlowTrigger({
-		storage: args.context.storage,
-		flowTriggerId: args.triggerId,
-	});
-	if (trigger === undefined || !trigger.enable) {
-		return null;
-	}
-
-	const workspace = await getWorkspace({
-		storage: args.context.storage,
-		experimental_storage: args.context.experimental_storage,
-		workspaceId: trigger.workspaceId,
-		useExperimentalStorage: args.useExperimentalStorage,
-	});
-
-	const triggerNode = workspace.nodes.find(
-		(node) => node.id === trigger.nodeId,
-	);
-	if (triggerNode === undefined) {
-		return null;
-	}
-
-	const workflow = buildWorkflowFromNode(triggerNode, workspace);
-	if (workflow === null) {
-		return null;
-	}
-
-	return {
-		workflow,
-		trigger,
-		workspace,
-	};
-}
 
 /** @todo telemetry */
 export async function createAndRunFlow(args: {
