+++ packages/giselle-engine/src/core/flows/index.ts
@@ -1,3 +1,4 @@
+export * from "./build-workflow-from-trigger";
 export * from "./configure-trigger";
 export * from "./create-and-run-flow";
 export * from "./create-run";
