+++ packages/giselle-engine/src/http/router.ts
@@ -311,6 +311,17 @@ export const createJsonRouters = {
 				return new Response(null, { status: 204 });
 			},
 		}),
+	buildWorkflowFromTrigger: (giselleEngine: GiselleEngine) =>
+		createHandler({
+			input: z.object({
+				triggerId: FlowTriggerId.schema,
+				useExperimentalStorage: z.boolean(),
+			}),
+			handler: async ({ input }) =>
+				JsonResponse.json({
+					result: await giselleEngine.buildWorkflowFromTrigger(input),
+				}),
+		}),
 	executeQuery: (giselleEngine: GiselleEngine) =>
 		createHandler({
 			input: z.object({
